Implemented fizzbuzz method
Loops all numbers 1-100
if divisible by both 3 and 5, it adds "FizzBuzz" to a string
if divisible by only 3, it adds "Fizz" to the string
if only divisible by 5, it adds "Buzz" to the string
if not divisible by either, it will add the number to the string
A space is added to the string between all values. At the end, the string is both printed and returned.